Split Binding.scala into {rx/mount}.scala
And two minor things:

- Change `.sequence` example from `Seq` to `List` to make it `cats` ready
- Rework some Scaladoc
